1.13 + 1.14 compatibility: the beginning.
- The plugin should be compatible with 1.13+. Compatibility with 1.12- has been dropped; use old versions for them.
- Map IDs are integer instead of shorts now, as the maps limits has been dropped by Minecraft.
- ImageOnMap items are now green to differentiate them. Map parts are lime.
- Removed glow on maps, as zLib's glow effect is not compatible with 1.13+.

+++ src/main/java/fr/moribus/imageonmap/image/ImageIOExecutor.java
@@ -60,12 +60,12 @@ public class ImageIOExecutor extends Worker
         });
     }
     
-    static public void saveImage(short mapID, BufferedImage image)
+    static public void saveImage(int mapID, BufferedImage image)
     {
         saveImage(ImageOnMap.getPlugin().getImageFile(mapID), image);
     }
     
-    static public void saveImage(short[] mapsIDs, PosterImage image)
+    static public void saveImage(int[] mapsIDs, PosterImage image)
     {
         for(int i = 0, c = mapsIDs.length; i < c; i++)
         {
@@ -75,7 +75,7 @@ public class ImageIOExecutor extends Worker
     
     static public void deleteImage(ImageMap map)
     {
-        short[] mapsIDs = map.getMapsIDs();
+        int[] mapsIDs = map.getMapsIDs();
         for(int i = 0, c = mapsIDs.length; i < c; i++)
         {
             deleteImage(ImageOnMap.getPlugin().getImageFile(mapsIDs[i]));
